keeper reordering
